- Improve the way that game sources are set.
Previously we get all files that is C/C++ files,
this create a problem when we have submodules that
depends of the same libraries.

For example, let's say that we have ProjectA that depends
on ProjectB and both of them depends of ProjectX.

ProjectA
|-----ProjectX
|-----ProjectB
      |-----ProjectX

This will cause compilation errors due the duplicated
definitions. So the correction is make the compile
ignore the sources from one of the ProjectX.

To achieve this we now have a separated script to update
the sources and a file to define all the paths that should
be ignored.